Add `hasBeenUsed` method to `AddressChange`
When an address change record is freshly initialized, it has the same
current and previous identifier. Each change to the record will change
the ID. This introduces the concept of a "used" record. This is
slightly different that the less granular concept of "isModified".

We'll use this in the Fundraising Application.