Keep identical address change IDs
An address change record allows two changes:
- setting an address
- opting out of the donation receipt

Each of those actions takes an `AddressChangeId` and "rotates" the
current ID to the `previousIdentifier`. When doing both changes, the Id
should only be rotated once, to avoid `hasBeenUsed` returning the wrong
value and to be able to check in the frontend if the previous ID still
returns a (used) address changed record.